Undo code move
